architecture MS 3

+++ src/main/java/io/eddvance/production/practice/translator_historique/repository/HistoriqueRepositoryInterface.java
@@ -3,5 +3,6 @@ package io.eddvance.production.practice.translator_historique.repository;
 import io.eddvance.production.practice.translator_historique.entity.historique.Historique;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface HistoriqueRepositoryInterface extends JpaRepository<Historique, Long> {
+
+public interface HistoriqueRepositoryInterface extends JpaRepository<Historique, Integer> {
 }
